ALE plots feature-wise

Makes an ALE plot for one single predictor variable of the model.

ALE_plot_split(object, variable, path_plot, nb_bins = 10)

Arguments

object
an object of class res_fitted_split

variable
environmental or genomic variable (PC derived from SNPs matrix or single SNP) for which the ALE plot should be calculated

path_plot
path where the ALE plot should be saved

nb_bins
number of intervals. Note that if nb_bins is provided by the user, the min_interval_points argument is ignored.

min_interval_points
minimum number of points contained within each interval (e.g. among the unique values taken by the feature). Values less than 4 should be avoided.

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de