Create a multi-environment trials data object, which will be used for prediction of unobserved data.

This function combines all types of predictor sources (genotypic, information about the environments to be predicted, environmental data if available...) and the data.frame describing the phenotypic observations to be predicted.

new_add_METData_to_predict(
  METData_training,
  geno_new = NULL,
  pheno_new = NULL,
  info_environments_to_predict = NULL,
  soil_variables = NULL,
  climate_variables = NULL,
  raw_weather_data = NULL,
  compute_climatic_ECs = FALSE,
  path_to_save = NULL,
  ...
)

add_METData_to_predict(
  METData_training,
  pheno_new = NULL,
  geno_new = NULL,
  compute_climatic_ECs = FALSE,
  info_environments_to_predict = NULL,
  climate_variables = NULL,
  soil_variables = NULL,
  raw_weather_data = NULL,
  path_to_save = NULL,
  ...
)

validate_add_METData_to_predict(x, ...)

Arguments

METData_training
METData object, which will be used as training set. Normally, it should be the same METData object as the one created with create_METData() and evaluated with predict_trait_MET_cv().

geno_new
data.frame object which should contain genotypic data for new candidates, if these are not included in METData_training$geno which will be used as training set. If already included, no need to include geno_new. Default is NULL.

pheno_new
data.frame object with at least 3 columns.

geno_ID character contains the genotype identifiers.

year numeric contains the year of the observation.

location character contains the name of the location.

Basically, this data.frame indicates which genotypes should be predicted in which environments. No columns corresponding to phentoypic traits is expected.

info_environments_to_predict
data.frame object with at least the 4 following columns.

year: numeric Year label of the environment

location: character Name of the location

longitude: numeric longitude of the environment

latitude: numeric latitude of the environment

The two next columns are required only if weather data should be retrieved from NASA POWER data using the argument compute_climatic_EC set to TRUE.

planting.date: (optional) Date YYYY-MM-DD

harvest.date: (optional) Date YYYY-MM-DD

The data.frame should contain as many rows as Year x Location combinations which will be used in pheno_new.

soil_variables
data.frame can be let as NULL by user, if no soil variables provided as input. Otherwise, a data.frame should be provided. The data.frame should contain as many rows as the info_environments_to_predict data.frame.
Columns should be:

year numeric with the year label

location character with the location character

Columns 3 and + should be numeric and contain the soil-based environmental covariates.

climate_variables
data.frame can be let as NULL by user, if no climate variables provided as input. Otherwise, a data.frame should be provided. The data.frame should contain as many rows as the info_environments_to_predict data.frame.
Columns should be:

year numeric with the year label

location character with the location character

Columns 3 and + should be numeric and contain the climate (weather-based) covariates.

If climate_variables is provided,compute_climatic_ECsshould be set to FALSE.

raw_weather_data
data.frame can be let as NULL by user, if no daily weather datasets are available. If else, required columns should be provided like this (colnames should be respected):

longitude numeric

latitude numeric

year numeric

location character

YYYYMMDD Date

Available weather data provided by user must be a subset of the following weather variable names. Colnames must be given as following:

T2M numeric Daily mean temperature (°C)

T2M_MIN numeric Daily minimum temperature (°C)

T2M_MAX numeric Daily maximum temperature (°C)

PRECTOTCORR numeric Daily total precipitation (mm)

RH2M numeric Daily mean relative humidity (%)

RH2M_MIN numeric Daily minimum relative humidity (%)

RH2M_MAX numeric Daily maximum relative humidity (%)

daily_solar_radiation numeric daily solar radiation (MJ/m^2/day)

top_atmosphere_insolation numeric Top-of-atmosphere Insolation (MJ/m^2/day)

T2MDEW numeric Dew Point (°C)

It is not required that weather data for ALL environments are provided by the user. If weather data for some environments are missing, they will be retrieved by the NASA

compute_climatic_ECs
logical indicates if climatic covariates should be computed with the function. Default is FALSE.
Set compute_climatic_ECs = TRUE if user wants to use weather data from NASA POWER data. For instance, if no weather-based covariables can be provided or if raw weather data are only available for some environments but not for others.

path_to_save
Path where daily weather data (if retrieved) and plots based on k-means clustering are saved.

Value

A formatted list of class METData which contains the following elements:

geno: matrix with genotype values of phenotyped individuals.

map: data.frame with genetic map.

pheno: data.frame with phenotypic trait values.

compute_climatic_ECs: logical indicates if environmental covariates were required to be retrieved via the package by the user.

env_data: data.frame with the environmental covariates per environment.

list_climatic_predictors: character with the names of the climatic predictor variables

list_soil_predictors: character with the names of the soil-based predictor variables

info_environments: data.frame contains basic information on each environment.

ECs_computed: OPTIONAL logical subelement added in the output only if the function get_ECs() was correctly run within the pipeline.

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de

Examples