Create a multi-environment trials data object

This function combines all types of data sources (genotypic, phenotypic, environmental data, information about the environments...) in a single data object of class METData.

new_add_new_METData(
  METData_training,
  pheno_new = NULL,
  geno_new = NULL,
  compute_climatic_ECs = FALSE,
  info_environments_to_predict = NULL,
  env_data_manual = NULL,
  raw_weather_data = NULL,
  ...
)

add_new_METData(
  METData_training,
  pheno_new = NULL,
  geno_new = NULL,
  compute_climatic_ECs = FALSE,
  info_environments_to_predict = NULL,
  env_data_manual = NULL,
  raw_weather_data = NULL,
  ...
)

validate_create_METData(x, ...)

Arguments

METData_training
Will be used as training set.

pheno_new
data.frame object with at least 3 columns.

geno_ID character contains the genotype identifiers.

year numeric contains the year of the observation.

location character contains the name of the location.

Basically, this data.frame only indicates which genotypes should be predicted in which environments. No columns with phenotypic values is expected.

info_environments_to_predict
data.frame object with at least the 4 first columns.

year: numeric Year label of the environment

location: character Name of the location

longitude: numeric longitude of the environment

latitude: numeric latitude of the environment

planting.date: (optional) Date YYYY-MM-DD

harvest.date: (optional) Date YYYY-MM-DD

The data.frame should contain as many rows as Year x Location combinations. Example: if only one location evaluated across four years, 4 rows should be present.

The fifth and sixth columns (planting.date and harvest.date) are required only if the user wants to download weather data with the package (setting argument ECs_computed = T in create_METData() and using subsequently the function get_ECs()).

scenario_weather_data
character. Options are:

manually: names of columns in the env_data_manual data.frame should be the same as in METData_training$env_data.

use_real_EC: environmental variables will be estimated for a set of "past" environments based on information provided in info_environments_to_predict. using the use_real_EC() function. This assumes that the environmental data can be retrieved and are from the past. only possible if... different according to compute_EC.

mean_previous_years: ECs computed based on average over 5 last years of the location indicated in info_environments_to_predict use of the location data from the location in the training set (implies that ) if computed_ECs== TRUE in METData_training. give fictive harvest date --> example

previous_year:

location_to_use
character. Default is NULL.

env_data
data.frame can be let as NULL by user, if no environment data provided as input. Otherwise, a data.frame should be provided. The data.frame should contain as many rows as the phenotypic dataset data.frame.
Columns should be:

geno_ID character with genotype ID

year numeric with the year label

location character with the location character

Columns 4 and + should be numeric and contain the environmental covariates.

Providing env_data and setting compute_climatic_ECs = T is possible. For instance, the user can have some information regarding the soil composition (\ A disease status can also be encoded as categorical variable if it affected some environments. In addition to these covariates, weather-based covariates will be computed if compute_climatic_ECs = T with the package if they were not available to provide as input by the user.

Value

A formatted list of class METData which contains the following elements:

geno: matrix with genotype values of phenotyped individuals.

map: data.frame with genetic map.

pheno: data.frame with phenotypic trait values.

compute_EC_by_geno: logical indicates if environmental covariates should be later computed.

env_data: data.frame with the environmental covariates per environment (and if genotype-specific, per genotype).

info_environments: data.frame contains basic information on each environment.

Author

Cathy C. Jubin cathy.jubin@uni-goettingen.de

Examples