Compute ECs based on day-windows of fixed length.

Compute the environmental covariates based on the daily weather table of an environment (Year x Location), and over day-windows which can be defined by the user for each environment (based for instance on observed phenological dates) in a table provided as input in create_METData().

compute_EC_user_defined_intervals(
  table_daily_W,
  intervals_growth_manual = NULL,
  base_temperature = 10,
  capped_max_temperature = F,
  max_temperature = 35,
  method_GDD_calculation = c("method_b"),
  ...
)

Arguments

table_daily_W
data.frame returned by the function get_daily_tables_per_env()

intervals_growth_manual
data.frame with:

column 1: numeric year

column 2: character location

columns 3 and +: numeric Date (in Days after Planting) at which the crop enters a new growth stage in a given environment. "P" refers to the planting date and should contain 0 as value, "VE" to emergence, etc... Day 0 (Planting Date, denoted "P") should be in the third column. At least 4 columns should be in this data.frame. There is no need to indicate the column "Harvest" - already considered in the function. An example of how this data.frame should be provided is given in intervals_growth_manual_G2F.

base_temperature
numeric Base temperature (crop growth assumed to be null below this value.) Default is 10.

method_GDD_calculation
character Method used to compute the GDD value, with one out of method_a or method_b.
method_a: No change of the value of \(T_{min}\). GDD = \(max (\frac{T_{min}+T_{max}}{2} - T_{base},0)\).
method_b: If \(T_{min}\) < \(T_{base}\), change \(T_{min}\) to \(T_{min}\) = \(T_{base}\).
Default = method_b.

Value

An object of class data.frame with 10 x number_total_fixed_windows + 1 last column (IDenv):

mean_TMIN: number_total_fixed_windows columns, indicating the average minimal temperature over the respective day-window.

mean_TMAX: number_total_fixed_windows columns, indicating the average maximal temperature over the respective day-window.

mean_TMEAN: number_total_fixed_windows columns, indicating the average mean temperature over the respective day-window.

freq_TMAX_sup30: number_total_fixed_windows columns, indicating the frequency of days with maximum temperature over 30°C over the respective day-window.

freq_TMAX_sup35: number_total_fixed_windows columns, indicating the frequency of days with maximum temperature over 35°C over the respective day-window.

sum_GDD: number_total_fixed_windows columns, indicating the growing degree days over the respective day-window.

sum_PTT: number_total_fixed_windows columns, indicating the accumulated photothermal time over the respective day-window.

sum_P: number_total_fixed_windows columns, indicating the accumulated precipitation over the respective day-window.

sum_et0: number_total_fixed_windows columns, indicating the cumulative reference evapotranspiration over the respective day-window.

freq_P_sup10: number_total_fixed_windows columns, indicating the frequency of days with total precipitation superior to 10 mm over the respective day-window.

sum_solar_radiation: number_total_fixed_windows columns, indicating the accumulated incoming solar radiation over the respective day-window.

mean_vapr_deficit: number_total_fixed_windows columns, indicating the mean vapour pressure deficit over the respective day-window.

IDenv character ID of the environment (Location_Year)

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de