S3 method used to fit an object of class rf_reg_1, rf_reg_2, rf_reg_3, xgb_reg_1, xgb_reg_2, xgb_reg_3,DL_reg,DL_reg_1, DL_reg_2,DL_reg_3,stacking_reg_1, stacking_reg_2 or stacking_reg_3.

S3 dispatching method for objects of class rf_reg_1, rf_reg_2, rf_reg_3, xgb_reg_1, xgb_reg_2, xgb_reg_3,DL_reg,DL_reg_1, DL_reg_2,DL_reg_3,stacking_reg_1, stacking_reg_2 or stacking_reg_3.

Fit a random forest model on an object of class rf_reg_1. Three hyperparameters (number of iterations = number of trees ; tree depth ; learning rate) are tuned using the training set via Bayesian optimization with 5-folds cross-validation (k-folds CV). A model is fitted on the training set using the best hyperparameters and model performance is evaluated on the test set.

Fit a random forest model on an object of class rf_reg_2. Three hyperparameters (number of iterations = number of trees ; tree depth ; learning rate) are tuned using the training set via Bayesian optimization with 5-folds cross-validation (k-folds CV). A model is fitted on the training set using the best hyperparameters and model performance is evaluated on the test set.

Fit a random forest model on an object of class rf_reg_3. Three hyperparameters (number of iterations = number of trees ; tree depth ; learning rate) are tuned using the training set via Bayesian optimization with 5-folds cross-validation (k-folds CV). A model is fitted on the training set using the best hyperparameters and model performance is evaluated on the test set.

Fit a gradient boosted trees model on an object of class xgb_reg_1. Three hyperparameters (number of iterations = number of trees ; tree depth ; learning rate) are tuned using the training set via Bayesian optimization with 5-folds cross-validation (k-folds CV). A model is fitted on the training set using the best hyperparameters and model performance is evaluated on the test set.

Fit a gradient boosted trees model on an object of class xgb_reg_2. Three hyperparameters (number of iterations = number of trees ; tree depth ; learning rate) are tuned using the training set via Bayesian optimization with 5-folds cross-validation (k-folds CV). A model is fitted on the training set using the best hyperparameters and model performance is evaluated on the test set.

Fit a gradient boosted trees model on an object of class xgb_reg_3. Three hyperparameters (number of iterations = number of trees ; tree depth ; learning rate) are tuned using the training set via Bayesian optimization with 5-folds cross-validation (k-folds CV). A model is fitted on the training set using the best hyperparameters and model performance is evaluated on the test set.

# S3 method for DL_reg_1
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 3,
  save_model = F,
  ...
)

# S3 method for DL_reg_2
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 3,
  save_model = F,
  ...
)

# S3 method for DL_reg_3
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 3,
  save_model = F,
  ...
)

fit_cv_split(object, ...)

# S3 method for default
fit_cv_split(object, ...)

# S3 method for rf_reg_1
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

# S3 method for rf_reg_2
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

# S3 method for rf_reg_3
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

# S3 method for stacking_reg_1
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  kernel_G = "linear",
  kernel_E = "polynomial",
  path_folder,
  save_model = F,
  ...
)

# S3 method for stacking_reg_2
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  kernel_G = "linear",
  kernel_E = "polynomial",
  kernel_GE = "polynomial",
  save_model = F,
  ...
)

# S3 method for stacking_reg_3
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 4,
  kernel_E = "polynomial",
  save_model = F,
  ...
)

# S3 method for xgb_reg_1
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

# S3 method for xgb_reg_2
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

# S3 method for xgb_reg_3
fit_cv_split(
  object,
  seed,
  inner_cv_reps = 1,
  inner_cv_folds = 5,
  save_model = F,
  ...
)

Arguments

object
an object of class xgb_reg_3

seed
integer Seed value.

inner_cv_reps
integer Number of repeats of the k-folds CV for hyperparameter optimization.

inner_cv_folds
integer Number k in the k-folds CV used for hyperparameter optimization.

Value

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de