Obtain daily climate data for an environment from NASA POWER data.

Function downloading daily weather data via the package nasapower based on longitude, latitude, planting and harvest date characterizing this environment.

get_daily_tables_per_env(environment, info_environments, ...)

Arguments

environment
character Name of the environment for which climate data should be extracted.

info_environments
data.frame object with at least the 4 first columns.

year: numeric Year label of the environment

location: character Name of the location

longitude: numeric longitude of the environment

latitude: numeric latitude of the environment

planting.date: (optional) Date YYYY-MM-DD

harvest.date: (optional) Date YYYY-MM-DD

IDenv: character ID of the environment (location x year)

The data.frame should contain as many rows as Year x Location combinations. Example: if only one location evaluated across four years, 4 rows should be present.

Value

a data.frame data.frame with the following columns extracted from POWER data, according to requested parameters:

longitude numeric

latitude numeric

YEAR numeric

MM integer

DD integer

DOY integer

YYYYMMDD Date

RH2M numeric

T2M numeric

T2M_MIN numeric

T2M_MAX numeric

PRECTOTCORR numeric

ALLSKY_SFC_SW_DWN numeric

T2MDEW numeric

IDenv character ID environment for which weather data were downloaded.

length.gs difftime length in days of the growing season for the environment.

References

Sparks AH (2018). “nasapower: a NASA POWER global meteorology, surface solar energy and climatology data client for R.” Journal of Open Source Software, 3 (30), 1035. Zotarelli L, Dukes MD, Romero CC, Migliaccio KW, Morgan KT (2010). “Step by step calculation of the Penman-Monteith Evapotranspiration (FAO-56 Method).” Institute of Food and Agricultural Sciences. University of Florida.

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de