Multi-year trial data of rice

This data are used as toy data for several functions of learnMET package.
Information about environments where japonica rice population was grown.

Format

A data.frame object with the 4 following columns.

year: numeric Year label of the environment

location: character Name of the location

longitude: numeric longitude of the environment

latitude: numeric latitude of the environment

Source

Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).

References

Monteverde E, Gutierrez L, Blanco P, Pérez de Vida F, Rosas JE, Bonnecarrère V, Quero G, McCouch S (2019). “Integrating molecular markers and environmental covariates to interpret genotype by environment interaction in rice (Oryza sativa L.) grown in subtropical areas.” G3: Genes, Genomes, Genetics, 9(5), 1519--1531.
Instituto Nacional de Investigación Agropecuaria (INIA). Programa Arroz. Estación Experimental INIA Treinta y Tres, Uruguay

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de