Compute marker effects per environment with Elastic Net

Compute marker effects per environment with Elastic Net

marker_effect_per_env_EN(
  geno,
  pheno,
  environment,
  pheno_trait,
  nb_folds_cv = 4,
  reps = 2,
  ...
)

Arguments

geno
data.frame with markers in columns and inviduals in rows. Typical input is METData$geno from the create_METData() function.

pheno
data.frame with: First column: ID genotypes Second column: year Third column: location Subsequent columns: phenotypic traits with names indicated in colnames() Last column: IDenv (combination LocationxYear) Typical input is METData$pheno from the create_METData() function.

environment
character indicating the name of the environment for which marker effects should be computed

pheno_trait
character Name of the trait under study for which marker effects should be estimated

nb_folds_cv
numeric Number of folds in the CV process

reps
numeric Number of repeats of the k-folds CV

Value

all_coef_avg data.frame

First column character contains the marker names.

Second column numeric the marker effects in this environment calculated by cross-validation with Elastic Net.

Third column character contains the environment name (combination LocationxYear).

Author

Cathy C. Jubin cathy.jubin@uni-goettingen.de