Compute marker P-values for each environment with GWAS.

GWAS method implemented: FarmCPU from GAPIT3. Multiple testing correction: Benjamini–Hochberg procedure with alpha=0.05.

marker_effect_per_env_FarmCPU(
  geno,
  pheno,
  map,
  environment,
  pheno_trait,
  nb_pcs = 5,
  ...
)

Arguments

geno
data.framewith markers in columns and inviduals in rows. Typical input is METData$geno from the create_METData() function.

pheno
data.frame object with at least 4 columns.

geno_ID character contains the genotype identifiers.

year numeric contains the year of the observation.

location character contains the name of the location.

From the fourth column on: each column numeric contains phenotypic values for a phenotypic trait observed in a combination Year x Location. Names of the traits can be provided as column names.

Last column: IDenv (combination LocationxYear)

Typical input is METData$pheno from the create_METData() function.

map
data.frame with 3 columns.

marker_name character with marker names

chr numeric with chromosome number

pos numeric with marker position.

Typical input is METData$map from the create_METData() function.

environment
character indicating the name of the environment for which marker effects should be computed

pheno_trait
character Name of the trait under study for which marker effects should be estimated

nb_pcs
numeric Number of principal components to use in FarmCPU analyses.

Value

a list which contains the following elements:

GWAS_results

data.frame FarmCPU results for all SNPs

threshold

numeric Cutoff derived from Benjamini-Hochberg procedure

selected_markers

character Vector containing the names of the SNPs passing the threshold

Details

For more information about the GWAS method, please consult the original publication from the authors: Liu, Xiaolei, et al. "Iterative usage of fixed and random effect models for powerful and efficient genome-wide association studies." PLoS genetics 12.2 (2016): e1005767. https://doi.org/10.1371/journal.pgen.1005767

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de