Multi-year trial data of rice

This data are used as toy data for several functions of learnMET package.
Phenotypic data for the indica rice population for four traits.

Format

data.frame object with 7 columns:

geno_ID character contains the genotype identifiers.

year numeric contains the year of the observation.

location character contains the name of the location.

GY character Grain Yield

PHR character Percentage of Head Rice Recovery

GC character Percentage of Chalky Grain

PH character Plant Height

Source

Data from the INIA's Rice Breeding Program (Uruguay). Data downloaded from publication Monteverde et al. (2019).

References

Monteverde E, Gutierrez L, Blanco P, P攼㸹rez de Vida F, Rosas JE, Bonnecarr攼㸸re V, Quero G, McCouch S (2019). “Integrating molecular markers and environmental covariates to interpret genotype by environment interaction in rice (Oryza sativa L.) grown in subtropical areas.” G3: Genes, Genomes, Genetics, 9(5), 1519--1531.
Instituto Nacional de Investigación Agropecuaria (INIA). Programa Arroz. Estación Experimental INIA Treinta y Tres, Uruguay

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de