Plot variable importance scores

Internal function of predict_trait_MET_cv().
This function plots variable importance scores if the compute_vip option is set to TRUE in predict_trait_MET_cv().
Plots are done at the CV scheme level from the , which means that:

If CV0 is evaluated, the plot shows the 40 most important variables according to the predicted element (i.e. site, year or environment).

For CV1 and CV2, variable importance plots are based on a average of the importance of each feature over all training/test splits.

Variable importance can be calculated based on model agnostic approaches (permutation-based methods, like for stacking_reg_1 or DL_reg), or on model-specific methods (gain metric for GBDT methods xgb_reg).

plot_results_vip_cv(
  fitting_all_splits,
  cv_type,
  cv0_type,
  path_folder,
  nb_folds_cv1,
  repeats_cv1,
  nb_folds_cv2,
  repeats_cv2
)

Arguments

fitting_all_splits
a list which is the list of results from the fitting step on all train/test partitions

cv_type
A character with one out of cv0 (prediction of new environments), cv00 (prediction of new genotypes in new environments), cv1 (prediction of new genotypes) or cv2 (prediction of incomplete field trials).

cv0_type
cv0_type A character with one out of leave-one-environment-out, leave-one-site-out,leave-one-year-out, forward-prediction.

path_folder
a character where the plots should be saved.

nb_folds_cv1
A numeric Number of folds used in the CV1 scheme.

repeats_cv1
A numeric Number of repeats in the CV1 scheme.

nb_folds_cv2
A numeric Number of folds used in the CV2 scheme.

repeats_cv2
A numeric Number of repeats in the CV2 scheme.

Value

A variable importance plot is saved in the path_folder. No specific object returned.

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de