Get train/test splits of the phenotypic MET dataset based on CV0.

Get train/test splits of the phenotypic MET dataset based on a number of random k-folds partitions determined by the user, according to the type CV00. Creation of the list of train/test splits based on phenotypic data, so that all the phenotypes from the same environment/year/site appear in the same fold, according to the type of the CV00 scheme. In addition to CV0 scheme, information on lines present in the test set evaluated in other environments are removed from the training set --> prediction of new genotypes in new environments.

predict_cv00(pheno_data, cv0_type)

Arguments

pheno_data
data.frame Dataset containing phenotypic outcome data, as well as the predictor variables.

cv0_type
character either leave-one-environment-out, leave-one-site-out, leave-one-year-out or forward-prediction.

Value

a cv_object object which contains the train/test splits of the CV scheme. Each element of the object corresponds to a split object with two elements:

training

data.frame Dataset with all observations for the training set.

test

data.frame Dataset with all observations for the test set.

References

Jarqu\'in D, Lemes da Silva C, Gaynor RC, Poland J, Fritz A, Howard R, Battenfield S, Crossa J (2017). “Increasing genomic-enabled prediction accuracy by modeling genotype\(\times\) environment interactions in Kansas wheat.” The plant genome, 10(2), 1--15. Jarqu\'in D, Crossa J, Lacaze X, Du Cheyron P, Daucourt J, Lorgeou J, Piraux F, Guerreiro L, Pérez P, Calus M, others (2014). “A reaction norm model for genomic selection using high-dimensional genomic and environmental data.” Theoretical and applied genetics, 127(3), 595--607.

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de