Phenotypic prediction of unobserved data.

Implement trait prediction based on SNP and environmental data with selection of prediction methods among Machine Learning approaches.

This function should be used to assess the predictive ability according to a cross-validation scheme determined by the user.

predict_trait_MET(
  METData_training,
  METData_new,
  trait,
  prediction_method = "xgb_reg",
  use_selected_markers = F,
  list_selected_markers_manual = NULL,
  geno_information = "PCs",
  lat_lon_included = F,
  year_included = F,
  include_env_predictors = T,
  list_env_predictors = NULL,
  seed = NULL,
  save_processing = T,
  path_folder,
  compute_vip = TRUE,
  ...
)

Arguments

trait
character Name of the trait to predict. An ordinal trait should be encoded as integer.

prediction_method
character specifying the predictive model to use. Options are xgb_reg (gradient boosted trees), (stacking of support vector machines with LASSO as meta-learner).

use_selected_markers
A Logical indicating whether to use a subset of markers obtained from a previous step (see function select_markers()).

geno_information
indicating how the complete genotypic matrix should be used in predictions. Options are SNPs (all of the markers will be individually used), PCs (PCA will be applied on each genotype matrix for the training set for dimensionality reduction) or PCs_G (decomposition of the genomic relationship matrix via PCA -- not yet implemented).

lat_lon_included
logical indicates if longitude and latitude data should be used as numeric predictors. Default is TRUE.

year_included
logical indicates if year factor should be used as predictor variable. Default is FALSE.

include_env_predictors
A logical indicating whether environmental covariates characterizing each environment should be used in predictions.

list_env_predictors
A character vector containing the names of the environmental predictors which should be used in predictions. By default NULL: all environmental predictors included in the env_data table of the METData object will be used.

seed
integer Seed value. Default is NULL. By default, a random seed will be generated.

save_processing
a logical indicating whether the processing steps obtained from the processing_train_test_split() or processing_train_test_split_kernel() functions should be saved in a .RDS object. Default is FALSE.

path_folder
a character indicating the full path where the .RDS object and plots generated during the analysis should be saved (do not use a Slash after the name of the last folder). Default is NULL.

...
Arguments passed to the processing_train_test_split(), processing_train_test_split_kernel(), reg_fitting_train_test_split(), reg_fitting_train_test_split_kernel() functions.

METData
list An object created by the initial function of the package create_METData().

cv_type
A character with one out of cv0 (prediction of new environments), cv00 (prediction of new genotypes in new environments), cv1 (prediction of new genotypes) or cv2 (prediction of incomplete field trials). Default is cv0.

cv0_type
A character with one out of leave-one-environment-out, leave-one-site-out,leave-one-year-out, forward-prediction. Default is leave-one-environment-out.

nb_folds_cv1
A numeric Number of folds used in the CV1 scheme. Default is 5.

repeats_cv1
A numeric Number of repeats in the CV1 scheme. Default is 50.

nb_folds_cv2
A numeric Number of folds used in the CV2 scheme. Default is 5.

repeats_cv2
A numeric Number of repeats in the CV2 scheme. Default is 50.

Value

A list object of class met_cv with the following items:

list_results_cv

list of res_fitted_split elements. Detailed prediction results for each split of the data within each element of this list.

seed_used

integer Seed used to generate the cross-validation splits.

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de