Quality control on daily weather data

This function checks range of values for METData and implements various test on daily weather data (persistence tests, internal consistency tests) provided by the user.

qc_raw_weather_data(daily_weather_data, info_environments, path_flagged_values)

Arguments

daily_weather_data
a data.frame which contains the following mandatory columns:

longitude numeric

latitude numeric

year numeric

location character

YYYYMMDD Date

IDenv character

DOY integer

Available weather data provided by user must be a subset of the following weather variable names. Column names of weather variables must be given as following: (For envs with raw daily weather data, T2M, T2M_MIN, T2M_MAX and PRECTOT are mandatory to provide and must be given without missing values, which implies that any imputation step should be performed before providing this dataset to the package. .):

T2M numeric Daily mean temperature (°C)

T2M_MIN numeric Daily minimum temperature (°C)

T2M_MAX numeric Daily maximum temperature (°C)

PRECTOT numeric Daily total precipitation (mm)

RH2M numeric Daily mean relative humidity (%)

RH2M_MIN numeric Daily minimum relative humidity (%)

RH2M_MAX numeric Daily maximum relative humidity (%)

daily_solar_radiation numeric daily solar radiation (MJ/m^2/day)

top_atmosphere_insolation numeric Top-of-atmosphere Insolation (MJ/m^2/day)

T2MDEW numeric Dew Point (°C)

Value

a processed data.frame after quality check with the same columns as before the QC.
Vapor pressure deficit is calculated if T2M_MIN, T2M_MAX, and either RH2M_MIN + RH2M_MAX or only RH2M are provided.
Warning messages are also thrown if some observations do not pass either the range test, persistence test or the internal consistency test. A data.frame with dubious values replaced by a "flagged" character is provided, along with explanations why these data were flagged (column "reason") is provided as output. None of the flagged values is assigned as missing values or transformed; therefore we strongly recommend the user to have a second look at the daily weather data provided and to correct potential dubious values indicated by the output of the present function.
Solar radiation or wind data are automatically retrieved from NASA, if they are not provided without any missing data by the user. As for any other weather variable used in this function, these data cannot be only partially provided (no missing values accepted).

References

Zotarelli L, Dukes MD, Romero CC, Migliaccio KW, Morgan KT (2010). “Step by step calculation of the Penman-Monteith Evapotranspiration (FAO-56 Method).” Institute of Food and Agricultural Sciences. University of Florida.

Author

Cathy C. Jubin cathy.jubin@uni-goettingen.de