Fitting a model on the training set and predicting the test set.

Function creating a workflow based on the prediction model chosen (all methods other than multi-kernel learning), with hyperparameter optimization using a Bayesian approach on the training set. Once the best hyperparameters are identified via resampling, the model is fitted on the complete dataset. Finally the test set is predicted.

reg_fitting_train_test_split(
  split,
  prediction_method = c("xgboost"),
  seed,
  inner_cv_reps = 2,
  inner_cv_folds = 5,
  return_finalized_train_test_sets = F,
  ...
)

Arguments

prediction_method
character Prediction method other than kernel-based methods to use on the predictors defined by the recipe item of the split object.Options are 'xgboost' for XGBoost, 'rf' for Random Forest. Default is xgboost.

seed
integer Seed value.

inner_cv_reps
integer Number of times to repeat the k-fold partitioning used for the inner cross-validation for estimation of the best hyperparameters. Default is 2.

inner_cv_folds
integer Number of partitions of the training set for the inner cross-validation for estimation of the best hyperparameters. Default is 5.

return_finalized_train_test_sets
a logical whether the trained dataset and the test dataset resulting from preprocessing operations should be returned in the final res_fitted_split object. Default is FALSE.

Value

a list object of class res_fitted_split with the following items:

predictions_df: data.frame with original test dataset with extra column containing predicted values.

cor_pred_obs: numeric Pearson's correlation between predicted and observed values of the test set.

rmse_pred_obs: numeric root mean square error between predicted and observed values of the test set.

best_hyperparameters: a tbl_df giving the tuning parameter combination with the best performance values which was used to fit the final model on the training set.

(optional if return_finalized_train_test_sets == TRUE) training: a data.frame with the train dataset after preprocessing steps

(optional if return_finalized_train_test_sets == TRUE) test: a data.frame with the test dataset after preprocessing steps

Author

Cathy C. Jubin cathy.jubin@uni-goettingen.de