Multiple kernel learning method based on G, E and GxE datasets.

A multiple kernel support vector machine framework using model stacking. Model stacking is an ensemble method that takes the outputs of different support vector machine models constructed based on different data types (genomic, environmental and GxE interactions). For each type of kernel, multiple model configurations are defined based on a grid of hyperparameters. Then, an ensemble is built with stacks to create an object that contain the assessment set predictions for each candidate ensemble member. A LASSO model is used to figure out how the respective output of each model from the stack members should be combined to obtain a final prediction, and to estimate the "stacking coefficients" of the model stack. Candidate members with a stacking coefficient different from 0 are trained on the full training set, and the test set can be predicted using the "instructions" on how to combine the respective predictions.

reg_fitting_train_test_split_kernel(
  split,
  seed,
  inner_cv_reps = 2,
  inner_cv_folds = 5,
  kernel_G = "rbf",
  kernel_E = "rbf",
  kernel_GE = "rbf",
  return_finalized_train_test_sets = F,
  ...
)

Arguments

split
An object of class split_processed object with the following items:

training: data.frame Training set.

test: data.frame Test set.

rec_G: recipe object with the different steps to process molecular marker dataset from the training set. Same transformations applied on the test set.

rec_E: recipe object with the different steps to process environmental data from the training set. Same transformations applied on the test set.

rec_GE: recipe object with the different steps to process the GxE dataset from the training set. Same transformations applied on the test set.

seed
integer Seed value.

inner_cv_reps
integer Number of times to repeat the k-fold partitioning used for the inner cross-validation for estimation of the best hyperparameters. The same resampling object is used for all kernel configurations which are evaluated. Default is 2.

inner_cv_folds
integer Number of partitions of the training set for the inner cross-validation for estimation of the best hyperparameters. The same resampling object is used for all kernel configurations which are evaluated. Default is 5.

kernel_G
character Type of kernel function to use for the molecular marker dataset. Options are rbf (default), polynomial or linear.

kernel_E
character Type of kernel function to use for the environmental dataset. Options are rbf (default), polynomial or linear.

kernel_GE
character Type of kernel function to use for the GxE dataset. Options are rbf (default), polynomial or linear.

return_finalized_train_test_sets
a logical whether the trained dataset and the test dataset resulting from preprocessing operations should be returned in the final res_fitted_split object. Default is FALSE.

Value

a list object of class res_fitted_split with the following items:

parameters_collection_G: a tbl_df mapping all of the candidate models based on genomic data to their hyperparameters with their stacking coefficient obtained after evaluation of the data stack on the training set.

parameters_collection_E: a tbl_df mapping all of the candidate models based on environmental data to their hyperparameters with their stacking coefficient.

parameters_collection_GE: a tbl_df mapping all of the candidate models based on GxE data to their hyperparameters with their stacking coefficient.

predictions_df: data.frame with original test dataset with extra column containing predicted values.

cor_pred_obs: numeric Pearson's correlation between predicted and observed values of the test set.

rmse_pred_obs: numeric root mean square error between predicted and observed values of the test set.

Details

For more information, consult: https://stacks.tidymodels.org/index.html

Author

Cathy C. Jubin cathy.jubin@uni-goettingen.de