Compute environmental covariates for each environment of the MET dataset.

This function enables to retrieve daily weather data from NASA POWER source for each environment and derive environmental covariates over non-overlapping time windows, which can be defined in various ways by the user. The user can also provide own daily weather data, even for only part of the total number of environments. For the remaining environments, weather data will be retrieved using the NASA POWER query.

get_ECs(
  info_environments,
  raw_weather_data = NULL,
  method_ECs_intervals = "fixed_nb_windows_across_env",
  length_minimum_gs = NULL,
  save_daily_weather_tables = T,
  path_data = NULL,
  crop_model = NULL,
  nb_windows_intervals = 10,
  duration_time_window_days = 10,
  base_temperature = 10,
  ...
)

Arguments

info_environments
a data.frame with the following columns

method_ECs_intervals
GDD

save_daily_weather_tables
logical indicates whether the daily weather tables should be saved. Default is TRUE.

path_data
character Path of the folder where a RDS object will be created to save the daily weather tables if saved. (Do not use a Slash after the name of the last folder.)

nb_windows_intervals
numeric Number of time windows covering the growing season length (common number of time windows across all environments). Default is 5.

...
Arguments passed to the compute_EC() function. Not all of the aforementioned weather variables need to be provided. Weather variables which are not provided and needed to compute environmental covariables will be retrieved from NASA POWER and merged to the weather data given by the user. If these environmental covariables should not be used in predictions, the user can specify in the prediction function the list of environmental covariables to use with the argument list_env_predictors. data are imputed if missing or assigned to NA after QC

fixed_length_time_windows_across_env
logical indicates if the growing season lengths should be divided in non-overlapping time windows of fixed lengths (in days) across all environments. This implies that the total number of time windows, which need to be common across all environments, is determined by the shortest growing season included in the MET environments. This further implies that the total growing season may not be covered by the environmental predictors for the longest growing seasons.
Default is TRUE.

fixed_nb_windows_across_env
logical indicates if the growing season lengths should be divided in a fixed number of non-overlapping windows, which fully cover the growing season of each environment. This means that these time windows might not be of same length across environments (but always of same length in one environment).
Default is FALSE.

customized_growth_intervals
logical indicates if the growth intervals to use are given by user and should be used to define time windows.
Default is FALSE.

Value

A data.frame object containing the weather-based environmental covariates.

References

Sparks AH (2018). “nasapower: a NASA POWER global meteorology, surface solar energy and climatology data client for R.” Journal of Open Source Software, 3 (30), 1035.

Author

Cathy C. Westhues cathy.jubin@uni-goettingen.de